requirements.txt
anthropic>=0.25.0
feedparser>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
